tools/token_server/requirements.txt
Flask==2.3.2
google-auth==2.20.0
